Rationalise layout of Execute/etc overloads
