--- src/Polly.Shared/ContextualPolicy.cs
@@ -1,412 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.Linq;
-using System.Threading;
-
-namespace Polly
-{
-    /// <summary>
-    /// Transient exception handling policies that can be applied to delegates.
-    /// </summary>
-    public partial class Policy
-    {
-        /// <summary>
-        /// Executes the specified action within the policy.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
-        [DebuggerStepThrough]
-        public void Execute(Action action, IDictionary<string, object> contextData)
-        {
-            Execute(ct => action(), new Context(contextData), CancellationToken.None);
-        }
-
-        /// <summary>
-        /// Executes the specified action within the policy.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="context">Context data that is passed to the exception policy.</param>
-        [DebuggerStepThrough]
-        public void Execute(Action action, Context context)
-        {
-            Execute(ct => action(), context, CancellationToken.None);
-        }
-
-        /// <summary>
-        /// Executes the specified action within the policy.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
-        /// <param name="cancellationToken">The cancellation token.</param>
-        /// <exception cref="System.ArgumentNullException">contextData</exception>
-        [DebuggerStepThrough]
-        public void Execute(Action<CancellationToken> action, IDictionary<string, object> contextData, CancellationToken cancellationToken)
-        {
-            Execute(action, new Context(contextData), cancellationToken);
-        }
-
-        /// <summary>
-        /// Executes the specified action within the policy.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="context">Context data that is passed to the exception policy.</param>
-        /// <param name="cancellationToken">The cancellation token.</param>
-        [DebuggerStepThrough]
-        public void Execute(Action<CancellationToken> action, Context context, CancellationToken cancellationToken)
-        {
-            if (_exceptionPolicy == null) throw new InvalidOperationException(
-                "Please use the synchronous-defined policies when calling the synchronous Execute (and similar) methods.");
-            if (context == null) throw new ArgumentNullException(nameof(context));
-
-            SetPolicyContext(context);
-
-            _exceptionPolicy(action, context, cancellationToken);
-        }
-
-        /// <summary>
-        /// Executes the specified action within the policy and returns the captured result.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
-        /// <exception cref="System.ArgumentNullException">contextData</exception>
-        /// <returns>The captured result</returns>
-        [DebuggerStepThrough]
-        public PolicyResult ExecuteAndCapture(Action action, IDictionary<string, object> contextData)
-        {
-            return ExecuteAndCapture(ct => action(), new Context(contextData), CancellationToken.None);
-        }
-
-        /// <summary>
-        /// Executes the specified action within the policy and returns the captured result.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="context">Context data that is passed to the exception policy.</param>
-        /// <returns>The captured result</returns>
-        [DebuggerStepThrough]
-        public PolicyResult ExecuteAndCapture(Action action, Context context)
-        {
-            return ExecuteAndCapture(ct => action(), context, CancellationToken.None);
-        }
-
-        /// <summary>
-        /// Executes the specified action within the policy and returns the captured result.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
-        /// <param name="cancellationToken">The cancellation token.</param>
-        /// <returns>The captured result</returns>
-        /// <exception cref="System.ArgumentNullException">contextData</exception>
-        [DebuggerStepThrough]
-        public PolicyResult ExecuteAndCapture(Action<CancellationToken> action, IDictionary<string, object> contextData, CancellationToken cancellationToken)
-        {
-            return ExecuteAndCapture(action, new Context(contextData), cancellationToken);
-        }
-
-        /// <summary>
-        /// Executes the specified action within the policy and returns the captured result
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="context">Context data that is passed to the exception policy.</param>
-        /// <param name="cancellationToken">The cancellation token.</param>
-        /// <returns>The captured result</returns>
-        [DebuggerStepThrough]
-        public PolicyResult ExecuteAndCapture(Action<CancellationToken> action, Context context, CancellationToken cancellationToken)
-        {
-            if (_exceptionPolicy == null) throw new InvalidOperationException(
-                "Please use the synchronous-defined policies when calling the synchronous Execute (and similar) methods.");
-            if (context == null) throw new ArgumentNullException(nameof(context));
-
-            try
-            {
-                Execute(action, context, cancellationToken);
-                return PolicyResult.Successful();
-            }
-            catch (Exception exception)
-            {
-                return PolicyResult.Failure(exception, GetExceptionType(_exceptionPredicates, exception));
-            }
-        }
-
-        /// <summary>
-        /// Executes the specified action within the policy and returns the result.
-        /// </summary>
-        /// <typeparam name="TResult">The type of the result.</typeparam>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
-        /// <exception cref="System.ArgumentNullException">contextData</exception>
-        /// <returns>
-        /// The value returned by the action
-        /// </returns>
-        /// <exception cref="System.ArgumentNullException">contextData</exception>
-        [DebuggerStepThrough]
-        public TResult Execute<TResult>(Func<TResult> action, IDictionary<string, object> contextData)
-        {
-            return Execute(ct => action(), new Context(contextData), CancellationToken.None);
-        }
-
-        /// <summary>
-        /// Executes the specified action within the policy and returns the result.
-        /// </summary>
-        /// <typeparam name="TResult">The type of the result.</typeparam>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="context">Context data that is passed to the exception policy.</param>
-        /// <exception cref="System.ArgumentNullException">contextData</exception>
-        /// <returns>
-        /// The value returned by the action
-        /// </returns>
-        /// <exception cref="System.ArgumentNullException">contextData</exception>
-        [DebuggerStepThrough]
-        public TResult Execute<TResult>(Func<TResult> action, Context context)
-        {
-            return Execute(ct => action(), context, CancellationToken.None);
-        }
-
-        /// <summary>
-        /// Executes the specified action within the policy and returns the result.
-        /// </summary>
-        /// <typeparam name="TResult">The type of the result.</typeparam>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
-        /// <param name="cancellationToken">The cancellation token.</param>
-        /// <returns>The value returned by the action</returns>
-        /// <exception cref="System.ArgumentNullException">contextData</exception>
-        [DebuggerStepThrough]
-        public TResult Execute<TResult>(Func<CancellationToken, TResult> action, IDictionary<string, object> contextData, CancellationToken cancellationToken)
-        {
-            return Execute(action, new Context(contextData), cancellationToken);
-        }
-
-        /// <summary>
-        /// Executes the specified action within the policy and returns the result.
-        /// </summary>
-        /// <typeparam name="TResult">The type of the result.</typeparam>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="context">Context data that is passed to the exception policy.</param>
-        /// <param name="cancellationToken">The cancellation token.</param>
-        /// <returns>The value returned by the action</returns>
-        [DebuggerStepThrough]
-        public TResult Execute<TResult>(Func<CancellationToken, TResult> action, Context context, CancellationToken cancellationToken)
-        {
-            if (_exceptionPolicy == null) throw new InvalidOperationException(
-                "Please use the synchronous-defined policies when calling the synchronous Execute (and similar) methods.");
-            if (context == null) throw new ArgumentNullException(nameof(context));
-
-            SetPolicyContext(context);
-
-            var result = default(TResult);
-            _exceptionPolicy(ct => { result = action(ct); }, context, cancellationToken);
-            return result;
-        }
-
-        /// <summary>
-        /// Executes the specified action within the policy and returns the captured result.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
-        /// <exception cref="System.ArgumentNullException">contextData</exception>
-        /// <returns>The captured result</returns>
-        [DebuggerStepThrough]
-        public PolicyResult<TResult> ExecuteAndCapture<TResult>(Func<TResult> action, IDictionary<string, object> contextData)
-        {
-            return ExecuteAndCapture(ct => action(), new Context(contextData), CancellationToken.None);
-        }
-
-
-        /// <summary>
-        /// Executes the specified action within the policy and returns the captured result.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="context">Context data that is passed to the exception policy.</param>
-        /// <exception cref="System.ArgumentNullException">contextData</exception>
-        /// <returns>The captured result</returns>
-        [DebuggerStepThrough]
-        public PolicyResult<TResult> ExecuteAndCapture<TResult>(Func<TResult> action, Context context)
-        {
-            return ExecuteAndCapture(ct => action(), context, CancellationToken.None);
-        }
-
-
-        /// <summary>
-        /// Executes the specified action within the policy and returns the captured result.
-        /// </summary>
-        /// <typeparam name="TResult">The type of the result.</typeparam>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
-        /// <param name="cancellationToken">The cancellation token.</param>
-        /// <returns>The captured result</returns>
-        /// <exception cref="System.ArgumentNullException">contextData</exception>
-        [DebuggerStepThrough]
-        public PolicyResult<TResult> ExecuteAndCapture<TResult>(Func<CancellationToken, TResult> action, IDictionary<string, object> contextData, CancellationToken cancellationToken)
-        {
-            return ExecuteAndCapture(action, new Context(contextData), cancellationToken);
-        }
-
-        /// <summary>
-        /// Executes the specified action within the policy and returns the captured result.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="context">Context data that is passed to the exception policy.</param>
-        /// <param name="cancellationToken">The cancellation token.</param>
-        /// <returns>The captured result</returns>
-        [DebuggerStepThrough]
-        public PolicyResult<TResult> ExecuteAndCapture<TResult>(Func<CancellationToken, TResult> action, Context context, CancellationToken cancellationToken)
-        {
-            if (_exceptionPolicy == null) throw new InvalidOperationException(
-                "Please use the synchronous-defined policies when calling the synchronous Execute (and similar) methods.");
-            if (context == null) throw new ArgumentNullException(nameof(context));
-
-            try
-            {
-                return PolicyResult<TResult>.Successful(Execute(action, context, cancellationToken));
-            }
-            catch (Exception exception)
-            {
-                return PolicyResult<TResult>.Failure(exception, GetExceptionType(_exceptionPredicates, exception));
-            }
-        }
-    }
-
-    /// <summary>
-    /// Transient fault handling policies that can be applied to delegates returning results of type <typeparamref name="TResult"/>
-    /// These policies can be called with arbitrary context data.
-    /// </summary>
-    public partial class Policy<TResult>
-    {
-        /// <summary>
-        /// Executes the specified action within the policy and returns the result.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
-        /// <exception cref="System.ArgumentNullException">contextData</exception>
-        /// <returns>
-        /// The value returned by the action
-        /// </returns>
-        /// <exception cref="System.ArgumentNullException">contextData</exception>
-        [DebuggerStepThrough]
-        public TResult Execute(Func<TResult> action, IDictionary<string, object> contextData)
-        {
-            return Execute(ct => action(), new Context(contextData), CancellationToken.None);
-        }
-
-        /// <summary>
-        /// Executes the specified action within the policy and returns the result.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="context">Context data that is passed to the exception policy.</param>
-        /// <exception cref="System.ArgumentNullException">context</exception>
-        /// <returns>
-        /// The value returned by the action
-        /// </returns>
-        /// <exception cref="System.ArgumentNullException">contextData</exception>
-        [DebuggerStepThrough]
-        public TResult Execute(Func<TResult> action, Context context)
-        {
-            return Execute(ct => action(), context, CancellationToken.None);
-        }
-
-        /// <summary>
-        /// Executes the specified action within the policy and returns the result.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
-        /// <param name="cancellationToken">The cancellation token.</param>
-        /// <returns>The value returned by the action</returns>
-        /// <exception cref="System.ArgumentNullException">contextData</exception>
-        [DebuggerStepThrough]
-        public TResult Execute(Func<CancellationToken, TResult> action, IDictionary<string, object> contextData, CancellationToken cancellationToken)
-        {
-            return Execute(action, new Context(contextData), cancellationToken);
-        }
-
-        /// <summary>
-        /// Executes the specified action within the policy and returns the result.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="context">Context data that is passed to the exception policy.</param>
-        /// <param name="cancellationToken">The cancellation token.</param>
-        /// <returns>The value returned by the action</returns>
-        [DebuggerStepThrough]
-        public TResult Execute(Func<CancellationToken, TResult> action, Context context, CancellationToken cancellationToken)
-        {
-            if (_executionPolicy == null) throw new InvalidOperationException(
-                "Please use the synchronous-defined policies when calling the synchronous Execute (and similar) methods.");
-            if (context == null) throw new ArgumentNullException(nameof(context));
-
-            SetPolicyContext(context);
-
-            return _executionPolicy(action, context, cancellationToken);
-        }
-
-        /// <summary>
-        /// Executes the specified action within the policy and returns the captured result.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
-        /// <exception cref="System.ArgumentNullException">contextData</exception>
-        /// <returns>The captured result</returns>
-        [DebuggerStepThrough]
-        public PolicyResult<TResult> ExecuteAndCapture(Func<TResult> action, IDictionary<string, object> contextData)
-        {
-            return ExecuteAndCapture(ct => action(), new Context(contextData), CancellationToken.None);
-        }
-
-        /// <summary>
-        /// Executes the specified action within the policy and returns the captured result.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="context">Context data that is passed to the exception policy.</param>
-        /// <exception cref="System.ArgumentNullException">contextData</exception>
-        /// <returns>The captured result</returns>
-        [DebuggerStepThrough]
-        public PolicyResult<TResult> ExecuteAndCapture(Func<TResult> action, Context context)
-        {
-            return ExecuteAndCapture(ct => action(), context, CancellationToken.None);
-        }
-
-        /// <summary>
-        /// Executes the specified action within the policy and returns the captured result.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
-        /// <param name="cancellationToken">The cancellation token.</param>
-        /// <returns>The captured result</returns>
-        /// <exception cref="System.ArgumentNullException">contextData</exception>
-        [DebuggerStepThrough]
-        public PolicyResult<TResult> ExecuteAndCapture(Func<CancellationToken, TResult> action, IDictionary<string, object> contextData, CancellationToken cancellationToken)
-        {
-            return ExecuteAndCapture(action, new Context(contextData), cancellationToken);
-        }
-
-        /// <summary>
-        /// Executes the specified action within the policy and returns the captured result.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="context">Context data that is passed to the exception policy.</param>
-        /// <param name="cancellationToken">The cancellation token.</param>
-        /// <returns>The captured result</returns>
-        [DebuggerStepThrough]
-        public PolicyResult<TResult> ExecuteAndCapture(Func<CancellationToken, TResult> action, Context context, CancellationToken cancellationToken)
-        {
-            if (_executionPolicy == null) throw new InvalidOperationException(
-                "Please use the synchronous-defined policies when calling the synchronous Execute (and similar) methods.");
-            if (context == null) throw new ArgumentNullException(nameof(context));
-
-            try
-            {
-                TResult result = Execute(action, context, cancellationToken);
-
-                if (_resultPredicates.Any(predicate => predicate(result)))
-                {
-                    return PolicyResult<TResult>.Failure(result);
-                }
-
-                return PolicyResult<TResult>.Successful(result);
-            }
-            catch (Exception exception)
-            {
-                return PolicyResult<TResult>.Failure(exception, GetExceptionType(_exceptionPredicates, exception));
-            }
-        }
-    }
-}
