--- src/Polly.Shared/ContextualPolicyAsync.cs
@@ -1,740 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.Linq;
-using System.Threading;
-using System.Threading.Tasks;
-
-namespace Polly
-{
-    public partial class Policy
-    {
-        /// <summary>
-        ///     Executes the specified asynchronous action within the policy.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
-        [DebuggerStepThrough]
-        public Task ExecuteAsync(Func<Task> action, IDictionary<string, object> contextData)
-        {
-            return ExecuteAsync(ct => action(), new Context(contextData), CancellationToken.None, false);
-        }
-
-        /// <summary>
-        ///     Executes the specified asynchronous action within the policy.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="context">Context data that is passed to the exception policy.</param>
-        [DebuggerStepThrough]
-        public Task ExecuteAsync(Func<Task> action, Context context)
-        {
-            return ExecuteAsync(ct => action(), context, CancellationToken.None, false);
-        }
-
-        /// <summary>
-        ///     Executes the specified asynchronous action within the policy.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
-        /// <param name="continueOnCapturedContext">Whether to continue on a captured synchronization context.</param>
-        [DebuggerStepThrough]
-        public Task ExecuteAsync(Func<Task> action, IDictionary<string, object> contextData, bool continueOnCapturedContext)
-        {
-            return ExecuteAsync(ct => action(), new Context(contextData), CancellationToken.None, continueOnCapturedContext);
-        }
-
-        /// <summary>
-        ///     Executes the specified asynchronous action within the policy.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="context">Context data that is passed to the exception policy.</param>
-        /// <param name="continueOnCapturedContext">Whether to continue on a captured synchronization context.</param>
-        [DebuggerStepThrough]
-        public Task ExecuteAsync(Func<Task> action, Context context, bool continueOnCapturedContext)
-        {
-            return ExecuteAsync(ct => action(), context, CancellationToken.None, continueOnCapturedContext);
-        }
-
-        /// <summary>
-        ///     Executes the specified asynchronous action within the policy.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
-        /// <param name="cancellationToken">A cancellation token which can be used to cancel the action.  When a retry policy in use, also cancels any further retries.</param>
-        [DebuggerStepThrough]
-        public Task ExecuteAsync(Func<CancellationToken, Task> action, IDictionary<string, object> contextData, CancellationToken cancellationToken)
-        {
-            return ExecuteAsync(action, new Context(contextData), cancellationToken, false);
-        }
-
-        /// <summary>
-        ///     Executes the specified asynchronous action within the policy.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="context">Context data that is passed to the exception policy.</param>
-        /// <param name="cancellationToken">A cancellation token which can be used to cancel the action.  When a retry policy in use, also cancels any further retries.</param>
-        [DebuggerStepThrough]
-        public Task ExecuteAsync(Func<CancellationToken, Task> action, Context context, CancellationToken cancellationToken)
-        {
-            return ExecuteAsync(action, context, cancellationToken, false);
-        }
-
-        /// <summary>
-        ///     Executes the specified asynchronous action within the policy.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
-        /// <param name="cancellationToken">A cancellation token which can be used to cancel the action.  When a retry policy in use, also cancels any further retries.</param>
-        /// <param name="continueOnCapturedContext">Whether to continue on a captured synchronization context.</param>
-        /// <exception cref="System.ArgumentNullException">contextData</exception>
-        [DebuggerStepThrough]
-        public Task ExecuteAsync(Func<CancellationToken, Task> action, IDictionary<string, object> contextData, CancellationToken cancellationToken, bool continueOnCapturedContext)
-        {
-            return ExecuteAsync(action, new Context(contextData), cancellationToken, continueOnCapturedContext);
-        }
-
-        /// <summary>
-        ///     Executes the specified asynchronous action within the policy.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="context">Context data that is passed to the exception policy.</param>
-        /// <param name="continueOnCapturedContext">Whether to continue on a captured synchronization context.</param>
-        /// <param name="cancellationToken">A cancellation token which can be used to cancel the action.  When a retry policy in use, also cancels any further retries.</param>
-        /// <exception cref="System.InvalidOperationException">Please use asynchronous-defined policies when calling asynchronous ExecuteAsync (and similar) methods.</exception>
-        [DebuggerStepThrough]
-        public async Task ExecuteAsync(Func<CancellationToken, Task> action, Context context, CancellationToken cancellationToken, bool continueOnCapturedContext)
-        {
-            if (_asyncExceptionPolicy == null) throw new InvalidOperationException
-                ("Please use asynchronous-defined policies when calling asynchronous ExecuteAsync (and similar) methods.");
-            if (context == null) throw new ArgumentNullException(nameof(context));
-
-            SetPolicyContext(context);
-
-            await _asyncExceptionPolicy(action, context, cancellationToken, continueOnCapturedContext).ConfigureAwait(continueOnCapturedContext);
-        }
-
-
-        /// <summary>
-        ///     Executes the specified asynchronous action within the policy and returns the captured result.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
-        /// <exception cref="System.ArgumentNullException">contextData</exception>
-        /// <returns>The captured result</returns>
-        [DebuggerStepThrough]
-        public Task<PolicyResult> ExecuteAndCaptureAsync(Func<Task> action, IDictionary<string, object> contextData)
-        {
-            return ExecuteAndCaptureAsync(action, new Context(contextData));
-        }
-
-        /// <summary>
-        ///     Executes the specified asynchronous action within the policy and returns the captured result.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="context">Context data that is passed to the exception policy.</param>
-        /// <returns>The captured result</returns>
-        [DebuggerStepThrough]
-        public Task<PolicyResult> ExecuteAndCaptureAsync(Func<Task> action, Context context)
-        {
-            return ExecuteAndCaptureAsync(action, context, false);
-        }
-
-        /// <summary>
-        ///     Executes the specified asynchronous action within the policy and returns the captured result.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
-        /// <param name="continueOnCapturedContext">Whether to continue on a captured synchronization context.</param>
-        /// <exception cref="System.ArgumentNullException">contextData</exception>
-        /// <returns>The captured result</returns>
-        [DebuggerStepThrough]
-        public Task<PolicyResult> ExecuteAndCaptureAsync(Func<Task> action, IDictionary<string, object> contextData, bool continueOnCapturedContext)
-        {
-            return ExecuteAndCaptureAsync(action, new Context(contextData), continueOnCapturedContext);
-        }
-
-        /// <summary>
-        ///     Executes the specified asynchronous action within the policy and returns the captured result.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="context">Context data that is passed to the exception policy.</param>
-        /// <param name="continueOnCapturedContext">Whether to continue on a captured synchronization context.</param>
-        /// <returns>The captured result</returns>
-        [DebuggerStepThrough]
-        public Task<PolicyResult> ExecuteAndCaptureAsync(Func<Task> action, Context context, bool continueOnCapturedContext)
-        {
-            return ExecuteAndCaptureAsync(ct => action(), context, CancellationToken.None, continueOnCapturedContext);
-        }
-
-        /// <summary>
-        ///     Executes the specified asynchronous action within the policy and returns the captured result.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
-        /// <param name="cancellationToken">A cancellation token which can be used to cancel the action.  When a retry policy in use, also cancels any further retries.</param>
-        /// <exception cref="System.ArgumentNullException">contextData</exception>
-        /// <returns>The captured result</returns>
-        [DebuggerStepThrough]
-        public Task<PolicyResult> ExecuteAndCaptureAsync(Func<CancellationToken, Task> action, IDictionary<string, object> contextData, CancellationToken cancellationToken)
-        {
-            return ExecuteAndCaptureAsync(action, new Context(contextData), cancellationToken);
-        }
-
-        /// <summary>
-        ///     Executes the specified asynchronous action within the policy and returns the captured result.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="context">Context data that is passed to the exception policy.</param>
-        /// <param name="cancellationToken">A cancellation token which can be used to cancel the action.  When a retry policy in use, also cancels any further retries.</param>
-        [DebuggerStepThrough]
-        public Task<PolicyResult> ExecuteAndCaptureAsync(Func<CancellationToken, Task> action, Context context, CancellationToken cancellationToken)
-        {
-            return ExecuteAndCaptureAsync(action, context, cancellationToken, false);
-        }
-
-        /// <summary>
-        ///     Executes the specified asynchronous action within the policy and returns the captured result.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
-        /// <param name="cancellationToken">A cancellation token which can be used to cancel the action.  When a retry policy in use, also cancels any further retries.</param>
-        /// <param name="continueOnCapturedContext">Whether to continue on a captured synchronization context.</param>
-        /// <exception cref="System.ArgumentNullException">contextData</exception>
-        /// <returns>The captured result</returns>
-        [DebuggerStepThrough]
-        public Task<PolicyResult> ExecuteAndCaptureAsync(Func<CancellationToken, Task> action, IDictionary<string, object> contextData, CancellationToken cancellationToken, bool continueOnCapturedContext)
-        {
-            return ExecuteAndCaptureAsync(action, new Context(contextData), cancellationToken,
-                continueOnCapturedContext);
-        }
-
-        /// <summary>
-        ///     Executes the specified asynchronous action within the policy and returns the captured result.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="context">Context data that is passed to the exception policy.</param>
-        /// <param name="continueOnCapturedContext">Whether to continue on a captured synchronization context.</param>
-        /// <param name="cancellationToken">A cancellation token which can be used to cancel the action.  When a retry policy in use, also cancels any further retries.</param>
-        /// <exception cref="System.InvalidOperationException">Please use asynchronous-defined policies when calling asynchronous ExecuteAsync (and similar) methods.</exception>
-        [DebuggerStepThrough]
-        public async Task<PolicyResult> ExecuteAndCaptureAsync(Func<CancellationToken, Task> action, Context context, CancellationToken cancellationToken, bool continueOnCapturedContext)
-        {
-            if (_asyncExceptionPolicy == null) throw new InvalidOperationException
-                ("Please use asynchronous-defined policies when calling asynchronous ExecuteAsync (and similar) methods.");
-            if (context == null) throw new ArgumentNullException(nameof(context));
-
-            try
-            {
-                await ExecuteAsync(action, context, cancellationToken, continueOnCapturedContext).ConfigureAwait(continueOnCapturedContext);
-                return PolicyResult.Successful();
-            }
-            catch (Exception exception)
-            {
-                return PolicyResult.Failure(exception, GetExceptionType(_exceptionPredicates, exception));
-            }
-
-        }
-
-        /// <summary>
-        ///     Executes the specified asynchronous action within the policy and returns the result.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
-        /// <returns>The value returned by the action</returns>
-        [DebuggerStepThrough]
-        public Task<TResult> ExecuteAsync<TResult>(Func<Task<TResult>> action, IDictionary<string, object> contextData)
-        {
-            return ExecuteAsync(ct => action(), new Context(contextData), CancellationToken.None, false);
-        }
-
-        /// <summary>
-        ///     Executes the specified asynchronous action within the policy and returns the result.
-        /// </summary>
-        /// <typeparam name="TResult">The type of the result.</typeparam>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="context">Context data that is passed to the exception policy.</param>
-        /// <returns>The value returned by the action</returns>
-        [DebuggerStepThrough]
-        public Task<TResult> ExecuteAsync<TResult>(Func<Task<TResult>> action, Context context)
-        {
-            return ExecuteAsync(ct => action(), context, CancellationToken.None, false);
-        }
-
-        /// <summary>
-        ///     Executes the specified asynchronous action within the policy and returns the result.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
-        /// <param name="continueOnCapturedContext">Whether to continue on a captured synchronization context.</param>
-        /// <returns>The value returned by the action</returns>
-        [DebuggerStepThrough]
-        public Task<TResult> ExecuteAsync<TResult>(Func<Task<TResult>> action, IDictionary<string, object> contextData, bool continueOnCapturedContext)
-        {
-            return ExecuteAsync(ct => action(), new Context(contextData), CancellationToken.None, continueOnCapturedContext);
-        }
-
-        /// <summary>
-        ///     Executes the specified asynchronous action within the policy and returns the result.
-        /// </summary>
-        /// <typeparam name="TResult">The type of the result.</typeparam>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="context">Context data that is passed to the exception policy.</param>
-        /// <param name="continueOnCapturedContext">Whether to continue on a captured synchronization context.</param>
-        /// <returns>The value returned by the action</returns>
-        [DebuggerStepThrough]
-        public Task<TResult> ExecuteAsync<TResult>(Func<Task<TResult>> action, Context context, bool continueOnCapturedContext)
-        {
-            return ExecuteAsync(ct => action(), context, CancellationToken.None, continueOnCapturedContext);
-        }
-        
-        /// <summary>
-        ///     Executes the specified asynchronous action within the policy and returns the result.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
-        /// <param name="cancellationToken">A cancellation token which can be used to cancel the action.  When a retry policy in use, also cancels any further retries.</param>
-        /// <returns>The value returned by the action</returns>
-        [DebuggerStepThrough]
-        public Task<TResult> ExecuteAsync<TResult>(Func<CancellationToken, Task<TResult>> action, IDictionary<string, object> contextData, CancellationToken cancellationToken)
-        {
-            return ExecuteAsync(action, new Context(contextData), cancellationToken, false);
-        }
-
-        /// <summary>
-        ///     Executes the specified asynchronous action within the policy and returns the result.
-        /// </summary>
-        /// <typeparam name="TResult">The type of the result.</typeparam>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="context">Context data that is passed to the exception policy.</param>
-        /// <param name="cancellationToken">A cancellation token which can be used to cancel the action.  When a retry policy is in use, also cancels any further retries.</param>
-        /// <returns>The value returned by the action</returns>
-        [DebuggerStepThrough]
-        public Task<TResult> ExecuteAsync<TResult>(Func<CancellationToken, Task<TResult>> action, Context context, CancellationToken cancellationToken)
-        {
-            return ExecuteAsync(action, context, cancellationToken, false);
-        }
-
-        /// <summary>
-        ///     Executes the specified asynchronous action within the policy and returns the result.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
-        /// <param name="cancellationToken">A cancellation token which can be used to cancel the action.  When a retry policy in use, also cancels any further retries.</param>
-        /// <param name="continueOnCapturedContext">Whether to continue on a captured synchronization context.</param>
-        /// <returns>The value returned by the action</returns>
-        /// <exception cref="System.ArgumentNullException">contextData</exception>
-        [DebuggerStepThrough]
-        public Task<TResult> ExecuteAsync<TResult>(Func<CancellationToken, Task<TResult>> action, IDictionary<string, object> contextData, CancellationToken cancellationToken, bool continueOnCapturedContext)
-        {
-            return ExecuteAsync(action, new Context(contextData), cancellationToken, continueOnCapturedContext);
-        }
-
-        /// <summary>
-        ///     Executes the specified asynchronous action within the policy and returns the result.
-        /// </summary>
-        /// <typeparam name="TResult">The type of the result.</typeparam>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="context">Context data that is passed to the exception policy.</param>
-        /// <param name="continueOnCapturedContext">Whether to continue on a captured synchronization context.</param>
-        /// <param name="cancellationToken">A cancellation token which can be used to cancel the action.  When a retry policy is in use, also cancels any further retries.</param>
-        /// <returns>The value returned by the action</returns>
-        /// <exception cref="System.InvalidOperationException">Please use asynchronous-defined policies when calling asynchronous ExecuteAsync (and similar) methods.</exception>
-        [DebuggerStepThrough]
-        public async Task<TResult> ExecuteAsync<TResult>(Func<CancellationToken, Task<TResult>> action, Context context, CancellationToken cancellationToken, bool continueOnCapturedContext)
-        {
-            if (_asyncExceptionPolicy == null) throw new InvalidOperationException(
-                "Please use asynchronous-defined policies when calling asynchronous ExecuteAsync (and similar) methods.");
-            if (context == null) throw new ArgumentNullException(nameof(context));
-
-            SetPolicyContext(context);
-
-            var result = default(TResult);
-            await _asyncExceptionPolicy(async ct =>
-            {
-                result = await action(ct).ConfigureAwait(continueOnCapturedContext);
-            }, context, cancellationToken, continueOnCapturedContext)
-            .ConfigureAwait(continueOnCapturedContext);
-            return result;
-        }
-
-        /// <summary>
-        /// Executes the specified asynchronous action within the policy and returns the result.
-        /// </summary>
-        /// <typeparam name="TResult">The type of the result.</typeparam>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
-        /// <exception cref="System.ArgumentNullException">contextData</exception>
-        /// <returns>The captured result</returns>
-        [DebuggerStepThrough]
-        public Task<PolicyResult<TResult>> ExecuteAndCaptureAsync<TResult>(Func<Task<TResult>> action, IDictionary<string, object> contextData)
-        {
-            return ExecuteAndCaptureAsync(action, new Context(contextData));
-        }
-
-        /// <summary>
-        /// Executes the specified asynchronous action within the policy and returns the result.
-        /// </summary>
-        /// <typeparam name="TResult">The type of the result.</typeparam>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="context">Context data that is passed to the exception policy.</param>
-        /// <returns>The captured result</returns>
-        [DebuggerStepThrough]
-        public Task<PolicyResult<TResult>> ExecuteAndCaptureAsync<TResult>(Func<Task<TResult>> action, Context context)
-        {
-            return ExecuteAndCaptureAsync(ct => action(), context, CancellationToken.None, false);
-        }
-
-        /// <summary>
-        /// Executes the specified asynchronous action within the policy and returns the result.
-        /// </summary>
-        /// <typeparam name="TResult">The type of the result.</typeparam>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
-        /// <param name="continueOnCapturedContext">Whether to continue on a captured synchronization context.</param>
-        /// <exception cref="System.ArgumentNullException">contextData</exception>
-        /// <returns>The captured result</returns>
-        [DebuggerStepThrough]
-        public Task<PolicyResult<TResult>> ExecuteAndCaptureAsync<TResult>(Func<Task<TResult>> action, IDictionary<string, object> contextData, bool continueOnCapturedContext)
-        {
-            return ExecuteAndCaptureAsync(action, new Context(contextData), continueOnCapturedContext);
-        }
-
-        /// <summary>
-        /// Executes the specified asynchronous action within the policy and returns the result.
-        /// </summary>
-        /// <typeparam name="TResult">The type of the result.</typeparam>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="context">Context data that is passed to the exception policy.</param>
-        /// <param name="continueOnCapturedContext">Whether to continue on a captured synchronization context.</param>
-        /// <returns>The captured result</returns>
-        [DebuggerStepThrough]
-        public Task<PolicyResult<TResult>> ExecuteAndCaptureAsync<TResult>(Func<Task<TResult>> action, Context context, bool continueOnCapturedContext)
-        {
-            return ExecuteAndCaptureAsync(ct => action(), context, CancellationToken.None, continueOnCapturedContext);
-        }
-
-        /// <summary>
-        ///     Executes the specified asynchronous action within the policy and returns the result.
-        /// </summary>
-        /// <typeparam name="TResult">The type of the result.</typeparam>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
-        /// <param name="cancellationToken">A cancellation token which can be used to cancel the action.  When a retry policy in use, also cancels any further retries.</param>
-        /// <exception cref="System.ArgumentNullException">contextData</exception>
-        /// <returns>The captured result</returns>
-        [DebuggerStepThrough]
-        public Task<PolicyResult<TResult>> ExecuteAndCaptureAsync<TResult>(Func<CancellationToken, Task<TResult>> action, IDictionary<string, object> contextData, CancellationToken cancellationToken)
-        {
-            return ExecuteAndCaptureAsync(action, new Context(contextData), cancellationToken, false);
-        }
-
-        /// <summary>
-        ///     Executes the specified asynchronous action within the policy and returns the result.
-        /// </summary>
-        /// <typeparam name="TResult">The type of the result.</typeparam>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="context">A cancellation token which can be used to cancel the action.  When a retry policy in use, also cancels any further retries.</param>
-        /// <param name="cancellationToken">A cancellation token which can be used to cancel the action.  When a retry policy in use, also cancels any further retries.</param>
-        /// <returns>The captured result</returns>
-        [DebuggerStepThrough]
-        public Task<PolicyResult<TResult>> ExecuteAndCaptureAsync<TResult>(Func<CancellationToken, Task<TResult>> action, Context context, CancellationToken cancellationToken)
-        {
-            return ExecuteAndCaptureAsync(action, context, cancellationToken, false);
-        }
-
-        /// <summary>
-        ///     Executes the specified asynchronous action within the policy and returns the result.
-        /// </summary>
-        /// <typeparam name="TResult">The type of the result.</typeparam>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
-        /// <param name="continueOnCapturedContext">Whether to continue on a captured synchronization context.</param>
-        /// <param name="cancellationToken">A cancellation token which can be used to cancel the action.  When a retry policy in use, also cancels any further retries.</param>
-        /// <returns>The captured result</returns>
-        /// <exception cref="System.ArgumentNullException">contextData</exception>
-        [DebuggerStepThrough]
-        public Task<PolicyResult<TResult>> ExecuteAndCaptureAsync<TResult>(Func<CancellationToken, Task<TResult>> action, IDictionary<string, object> contextData, CancellationToken cancellationToken, bool continueOnCapturedContext)
-        {
-            return ExecuteAndCaptureAsync(action, new Context(contextData), cancellationToken, continueOnCapturedContext);
-        }
-
-        /// <summary>
-        ///     Executes the specified asynchronous action within the policy and returns the result.
-        /// </summary>
-        /// <typeparam name="TResult">The type of the result.</typeparam>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="context">Context data that is passed to the exception policy.</param>
-        /// <param name="cancellationToken">A cancellation token which can be used to cancel the action.  When a retry policy in use, also cancels any further retries.</param>
-        /// <param name="continueOnCapturedContext">Whether to continue on a captured synchronization context.</param>
-        /// <returns>The captured result</returns>
-        /// <exception cref="System.InvalidOperationException">Please use asynchronous-defined policies when calling asynchronous ExecuteAsync (and similar) methods.</exception>
-        [DebuggerStepThrough]
-        public async Task<PolicyResult<TResult>> ExecuteAndCaptureAsync<TResult>(Func<CancellationToken, Task<TResult>> action, Context context, CancellationToken cancellationToken, bool continueOnCapturedContext)
-        {
-            if (_asyncExceptionPolicy == null) throw new InvalidOperationException(
-                "Please use asynchronous-defined policies when calling asynchronous ExecuteAsync (and similar) methods.");
-            if (context == null) throw new ArgumentNullException(nameof(context));
-
-            try
-            {
-                return PolicyResult<TResult>.Successful(await ExecuteAsync(action, context, cancellationToken, continueOnCapturedContext).ConfigureAwait(continueOnCapturedContext));
-            }
-            catch (Exception exception)
-            {
-                return PolicyResult<TResult>.Failure(exception, GetExceptionType(_exceptionPredicates, exception));
-            }
-        }
-    }
-
-    public partial class Policy<TResult>
-    {
-        /// <summary>
-        ///     Executes the specified asynchronous action within the policy and returns the result.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
-        /// <returns>The value returned by the action</returns>
-        [DebuggerStepThrough]
-        public Task<TResult> ExecuteAsync(Func<Task<TResult>> action, IDictionary<string, object> contextData)
-        {
-            return ExecuteAsync(ct => action(), new Context(contextData), CancellationToken.None, false);
-        }
-
-        /// <summary>
-        ///     Executes the specified asynchronous action within the policy and returns the result.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="context">Context data that is passed to the exception policy.</param>
-        /// <returns>The value returned by the action</returns>
-        [DebuggerStepThrough]
-        public Task<TResult> ExecuteAsync(Func<Task<TResult>> action, Context context)
-        {
-            return ExecuteAsync(ct => action(), context, CancellationToken.None, false);
-        }
-
-        /// <summary>
-        ///     Executes the specified asynchronous action within the policy and returns the result.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
-        /// <param name="continueOnCapturedContext">Whether to continue on a captured synchronization context.</param>
-        /// <returns>The value returned by the action</returns>
-        [DebuggerStepThrough]
-        public Task<TResult> ExecuteAsync(Func<Task<TResult>> action, IDictionary<string, object> contextData, bool continueOnCapturedContext)
-        {
-            return ExecuteAsync(ct => action(), new Context(contextData), CancellationToken.None, continueOnCapturedContext);
-        }
-
-        /// <summary>
-        ///     Executes the specified asynchronous action within the policy and returns the result.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="context">Context data that is passed to the exception policy.</param>
-        /// <param name="continueOnCapturedContext">Whether to continue on a captured synchronization context.</param>
-        /// <returns>The value returned by the action</returns>
-        [DebuggerStepThrough]
-        public Task<TResult> ExecuteAsync(Func<Task<TResult>> action, Context context, bool continueOnCapturedContext)
-        {
-            return ExecuteAsync(ct => action(), context, CancellationToken.None, continueOnCapturedContext);
-        }
-
-        /// <summary>
-        ///     Executes the specified asynchronous action within the policy and returns the result.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
-        /// <param name="cancellationToken">A cancellation token which can be used to cancel the action.  When a retry policy in use, also cancels any further retries.</param>
-        /// <returns>The value returned by the action</returns>
-        [DebuggerStepThrough]
-        public Task<TResult> ExecuteAsync(Func<CancellationToken, Task<TResult>> action, IDictionary<string, object> contextData, CancellationToken cancellationToken)
-        {
-            return ExecuteAsync(action, new Context(contextData), cancellationToken, false);
-        }
-
-        /// <summary>
-        ///     Executes the specified asynchronous action within the policy and returns the result.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="context">Context data that is passed to the exception policy.</param>
-        /// <param name="cancellationToken">A cancellation token which can be used to cancel the action.  When a retry policy is in use, also cancels any further retries.</param>
-        /// <returns>The value returned by the action</returns>
-        [DebuggerStepThrough]
-        public Task<TResult> ExecuteAsync(Func<CancellationToken, Task<TResult>> action, Context context, CancellationToken cancellationToken)
-        {
-            return ExecuteAsync(action, context, cancellationToken, false);
-        }
-
-        /// <summary>
-        ///     Executes the specified asynchronous action within the policy and returns the result.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
-        /// <param name="cancellationToken">A cancellation token which can be used to cancel the action.  When a retry policy in use, also cancels any further retries.</param>
-        /// <param name="continueOnCapturedContext">Whether to continue on a captured synchronization context.</param>
-        /// <returns>The value returned by the action</returns>
-        /// <exception cref="System.ArgumentNullException">contextData</exception>
-        [DebuggerStepThrough]
-        public Task<TResult> ExecuteAsync(Func<CancellationToken, Task<TResult>> action, IDictionary<string, object> contextData, CancellationToken cancellationToken, bool continueOnCapturedContext)
-        {
-            return ExecuteAsync(action, new Context(contextData), cancellationToken, continueOnCapturedContext);
-        }
-
-
-        /// <summary>
-        ///     Executes the specified asynchronous action within the policy and returns the result.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="context">Context data that is passed to the exception policy.</param>
-        /// <param name="continueOnCapturedContext">Whether to continue on a captured synchronization context.</param>
-        /// <param name="cancellationToken">A cancellation token which can be used to cancel the action.  When a retry policy is in use, also cancels any further retries.</param>
-        /// <returns>The value returned by the action</returns>
-        /// <exception cref="System.InvalidOperationException">Please use asynchronous-defined policies when calling asynchronous ExecuteAsync (and similar) methods.</exception>
-        [DebuggerStepThrough]
-        public async Task<TResult> ExecuteAsync(Func<CancellationToken, Task<TResult>> action, Context context, CancellationToken cancellationToken, bool continueOnCapturedContext)
-        {
-            if (_asyncExecutionPolicy == null) throw new InvalidOperationException(
-                "Please use asynchronous-defined policies when calling asynchronous ExecuteAsync (and similar) methods.");
-            if (context == null) throw new ArgumentNullException(nameof(context));
-
-            SetPolicyContext(context);
-
-            TResult result = await _asyncExecutionPolicy(
-                action,
-                context,
-                cancellationToken,
-                continueOnCapturedContext)
-            .ConfigureAwait(continueOnCapturedContext);
-            return result;
-        }
-
-        /// <summary>
-        /// Executes the specified asynchronous action within the policy and returns the result.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
-        /// <exception cref="System.ArgumentNullException">contextData</exception>
-        /// <returns>The captured result</returns>
-        [DebuggerStepThrough]
-        public Task<PolicyResult<TResult>> ExecuteAndCaptureAsync(Func<Task<TResult>> action, IDictionary<string, object> contextData)
-        {
-            return ExecuteAndCaptureAsync(ct => action(), new Context(contextData), CancellationToken.None, false);
-        }
-
-        /// <summary>
-        /// Executes the specified asynchronous action within the policy and returns the result.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="context">Context data that is passed to the exception policy.</param>
-        /// <returns>The captured result</returns>
-        [DebuggerStepThrough]
-        public Task<PolicyResult<TResult>> ExecuteAndCaptureAsync(Func<Task<TResult>> action, Context context)
-        {
-            return ExecuteAndCaptureAsync(ct => action(), context, CancellationToken.None, false);
-        }
-
-        /// <summary>
-        /// Executes the specified asynchronous action within the policy and returns the result.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
-        /// <param name="continueOnCapturedContext">Whether to continue on a captured synchronization context.</param>
-        /// <exception cref="System.ArgumentNullException">contextData</exception>
-        /// <returns>The captured result</returns>
-        [DebuggerStepThrough]
-        public Task<PolicyResult<TResult>> ExecuteAndCaptureAsync(Func<Task<TResult>> action, IDictionary<string, object> contextData, bool continueOnCapturedContext)
-        {
-            return ExecuteAndCaptureAsync(ct => action(), new Context(contextData), CancellationToken.None, continueOnCapturedContext);
-        }
-
-
-        /// <summary>
-        /// Executes the specified asynchronous action within the policy and returns the result.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="context">Context data that is passed to the exception policy.</param>
-        /// <param name="continueOnCapturedContext">Whether to continue on a captured synchronization context.</param>
-        /// <returns>The captured result</returns>
-        [DebuggerStepThrough]
-        public Task<PolicyResult<TResult>> ExecuteAndCaptureAsync(Func<Task<TResult>> action, Context context, bool continueOnCapturedContext)
-        {
-            return ExecuteAndCaptureAsync(ct => action(), context, CancellationToken.None, continueOnCapturedContext);
-        }
-
-        /// <summary>
-        ///     Executes the specified asynchronous action within the policy and returns the result.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
-        /// <param name="cancellationToken">A cancellation token which can be used to cancel the action.  When a retry policy in use, also cancels any further retries.</param>
-        /// <exception cref="System.ArgumentNullException">contextData</exception>
-        /// <returns>The captured result</returns>
-        [DebuggerStepThrough]
-        public Task<PolicyResult<TResult>> ExecuteAndCaptureAsync(Func<CancellationToken, Task<TResult>> action, IDictionary<string, object> contextData, CancellationToken cancellationToken)
-        {
-            return ExecuteAndCaptureAsync(action, new Context(contextData), cancellationToken, false);
-        }
-
-        /// <summary>
-        ///     Executes the specified asynchronous action within the policy and returns the result.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="context">Context data that is passed to the exception policy.</param>
-        /// <param name="cancellationToken">A cancellation token which can be used to cancel the action.  When a retry policy in use, also cancels any further retries.</param>
-        /// <returns>The captured result</returns>
-        [DebuggerStepThrough]
-        public Task<PolicyResult<TResult>> ExecuteAndCaptureAsync(Func<CancellationToken, Task<TResult>> action, Context context, CancellationToken cancellationToken)
-        {
-            return ExecuteAndCaptureAsync(action, context, cancellationToken, false);
-        }
-
-        /// <summary>
-        ///     Executes the specified asynchronous action within the policy and returns the result.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
-        /// <param name="continueOnCapturedContext">Whether to continue on a captured synchronization context.</param>
-        /// <param name="cancellationToken">A cancellation token which can be used to cancel the action.  When a retry policy in use, also cancels any further retries.</param>
-        /// <returns>The captured result</returns>
-        /// <exception cref="System.ArgumentNullException">contextData</exception>
-        [DebuggerStepThrough]
-        public Task<PolicyResult<TResult>> ExecuteAndCaptureAsync(Func<CancellationToken, Task<TResult>> action, IDictionary<string, object> contextData, CancellationToken cancellationToken, bool continueOnCapturedContext)
-        {
-            return ExecuteAndCaptureAsync(action, new Context(contextData), cancellationToken, continueOnCapturedContext);
-        }
-
-        /// <summary>
-        ///     Executes the specified asynchronous action within the policy and returns the result.
-        /// </summary>
-        /// <param name="action">The action to perform.</param>
-        /// <param name="context">Context data that is passed to the exception policy.</param>
-        /// <param name="continueOnCapturedContext">Whether to continue on a captured synchronization context.</param>
-        /// <param name="cancellationToken">A cancellation token which can be used to cancel the action.  When a retry policy in use, also cancels any further retries.</param>
-        /// <returns>The captured result</returns>
-        /// <exception cref="System.InvalidOperationException">Please use asynchronous-defined policies when calling asynchronous ExecuteAsync (and similar) methods.</exception>
-        [DebuggerStepThrough]
-        public async Task<PolicyResult<TResult>> ExecuteAndCaptureAsync(Func<CancellationToken, Task<TResult>> action, Context context, CancellationToken cancellationToken, bool continueOnCapturedContext)
-        {
-            if (_asyncExecutionPolicy == null) throw new InvalidOperationException(
-                "Please use asynchronous-defined policies when calling asynchronous ExecuteAsync (and similar) methods.");
-            if (context == null) throw new ArgumentNullException(nameof(context));
-
-            try
-            {
-                TResult result = await ExecuteAsync(action, context, cancellationToken, continueOnCapturedContext).ConfigureAwait(continueOnCapturedContext);
-
-                if (_resultPredicates.Any(predicate => predicate(result)))
-                {
-                    return PolicyResult<TResult>.Failure(result);
-                }
-
-                return PolicyResult<TResult>.Successful(result);
-            }
-            catch (Exception exception)
-            {
-                return PolicyResult<TResult>.Failure(exception, GetExceptionType(_exceptionPredicates, exception));
-            }
-        }
-    }
-
-}
-
